backend/requirements.txt
### Dependencias Python

# ============================================
# requirements.txt - Backend Dependencies
# Plataforma Integral de Rendimiento Estudiantil
# ============================================

# ===== Core Framework =====
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3

# ===== Database =====
SQLAlchemy==2.0.43
mysql-connector-python==9.4.0
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# ===== Document Processing =====
PyPDF2==3.0.1
#PyMuPDF==1.24.0
python-docx==1.1.0
Pillow==11.3.0

# ===== Computer Vision & Video =====
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
numpy==2.2.6

# ===== Deep Learning & AI =====
deepface==0.0.95
tensorflow==2.20.0
tf-keras==2.20.1
mtcnn==1.0.0
retina-face==0.0.17

# ===== NLP Processing =====
spacy==3.7.2
nltk==3.8.1

# ===== Audio Processing =====
SpeechRecognition==3.13.0
pydub==0.25.1
PyAudio==0.2.14

# ===== Google Gemini API =====
google-generativeai==0.4.6

# ===== Office Document Generation =====
python-pptx==0.6.23
openpyxl==3.1.2

# ===== Utilities =====
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.6
MarkupSafe==3.0.3

# ===== Security & Auth =====
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# ===== Testing =====
pytest==8.4.2
pytest-cov==7.0.0
pytest-flask==1.3.0
coverage==7.10.7

# ===== Development Tools =====
black==24.1.1
flake8==7.0.0
autopep8==2.0.4

# ===== Monitoring & Logging =====
colorlog==6.8.2
python-json-logger==2.0.7

# ===== File Handling =====
python-magic==0.4.27
watchdog==4.0.0

# ===== Date & Time =====
python-dateutil==2.9.0.post0
pytz==2025.2

# ===== HTTP & API =====
httpx==0.26.0
aiohttp==3.9.1

# ===== Data Processing =====
pandas==2.3.3
scipy==1.12.0

# ===== Background Tasks (Optional) =====
celery==5.3.6
redis==5.0.1

# ===== Production Server =====
gunicorn==23.0.0
gevent==24.2.1